Interim fix for local blocklist memory issue
Issue fix: After a few VPN restarts with local blocklists enabled, identified a memory
leak because of allocation made for local blocklists.

Now the BraveDNS object is created once instead of recreating on every restart
